Adding SSL private key encryption support
Currently when using an SSL key the encryption key must be stored in clear text. This requires the SSL key to be either stored within the repository or stored outside of it.

This is fine for internal SSL certificates (or more specifically their private keys) where their exposure wouldn't cause the end of the world.

Looking forwards we will want to use 'real' keys where their exposure would be a problem. With Stash we could simply store such keys privately - but this isn't particularly secure (unless we heavily restricted access to repositories and had complete faith in all users with administrative privileges).

One solution to this problem is to encrypt the private key then decrypt it for use. This would allow the key to be stored in repositories but require an extra secret (the pass phrase) in order to use the key. How the pass-phrase is stored is a separate problem, with multiple solutions depending on scale/security etc.

To support encrypted private keys new a task has been added to read in an encrypted key and return a decrypted version using OpenSSL. Variables are added to specify the pass-phrase (which should be passed at run-time).

This approach is based on this article/guide:

http://red-badger.com/blog/2014/02/28/deploying-ssl-keys-securely-with-ansible/